requirements.txt
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
fastapi==0.100.0
frozenlist==1.4.0
gunicorn==21.2.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
multidict==6.0.4
numpy==1.23.5
opencv-contrib-python-headless~=4.7.0.72
packaging==23.1
Pillow==9.5.0
pydantic==1.10.7
PyMuPDF==1.22.5
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
ujson==5.8.0
uvicorn==0.23.1
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2
setuptools==68.0.0
pathos~=0.3.1
uvloop==0.17.0

boto3==1.28.17
aioboto3~=11.3.0
botocore~=1.31.17